requirements.txt
requests>=2.28
python-dotenv>=1.0
pytest>=7.0
google-generativeai>=0.8.0
flask>=2.0
pytest-mock>=3.10
sqlalchemy>=1.4
psycopg2-binary>=2.9
pymysql>=1.0
google-api-python-client>=2.0
google-auth-httplib2>=0.1
google-auth-oauthlib>=0.5
